Refactor products view
