+++ src/views/Products/components/TickerCell.tsx
